+++ sdk/core/typespec_client_core/src/http/clients/mod.rs
@@ -35,13 +35,13 @@ pub fn new_http_client() -> Arc<dyn HttpClient> {
 pub trait HttpClient: Send + Sync + std::fmt::Debug {
     /// Send a request to the service.
     ///
-    /// It does not consume the request. Implementors are expected to clone the necessary parts
-    /// of the request and pass them to the underlying transport.
+    /// Implementors should consume the body from the request by calling
+    /// [`Body::take`](crate::http::request::Body::take) to gain ownership.
     ///
     /// # Errors
     ///
     /// The built-in [`RetryPolicy`](crate::http::policies::RetryPolicy) will resend the [`Request`]
     /// for some [`ErrorKind::HttpResponse`](crate::error::ErrorKind::HttpResponse) status codes e.g., [`StatusCode::TooManyRequests`](crate::http::StatusCode::TooManyRequests) and
     /// for [`ErrorKind::Io`](crate::error::ErrorKind::Io) returned by your `HttpClient` for situations like connection resets.
-    async fn execute_request(&self, request: &Request) -> Result<AsyncRawResponse>;
+    async fn execute_request(&self, request: &mut Request) -> Result<AsyncRawResponse>;
 }