+++ sdk/core/typespec_client_core/src/http/clients/noop.rs
@@ -17,7 +17,7 @@ pub(crate) fn new_noop_client() -> std::sync::Arc<dyn super::HttpClient> {
 #[async_trait]
 impl super::HttpClient for NoopClient {
     #[allow(clippy::diverging_sub_expression)]
-    async fn execute_request(&self, request: &Request) -> Result<AsyncRawResponse> {
+    async fn execute_request(&self, request: &mut Request) -> Result<AsyncRawResponse> {
         panic!(
             "A request was called on the default http client `NoopClient`.\
 	This client does nothing but panic. Make sure to enable an http\